backwards-incompatible api changes for 2020-03-02, bringing us forward from 2018-02-28

+++ lib/fabric/app/models/fabric/coupon.rb
@@ -15,7 +15,7 @@ module Fabric
     field :max_redemptions, type: Integer
     field :metadata, type: Hash
     field :name, type: String
-    field :percent_off, type: Integer
+    field :percent_off, type: Float
     field :redeem_by, type: Time
     field :times_redeemed, type: Integer
     field :coupon_valid, type: Boolean
